+++ common/src/main/java/com/binaris/wizardry/content/spell/healing/ArcaneJammer.java
@@ -38,7 +38,7 @@ public class ArcaneJammer extends RaySpell {
         if (entityHit.getEntity() instanceof LivingEntity target) {
             if (!ctx.world().isClientSide) {
                 target.addEffect(new MobEffectInstance(EBMobEffects.ARCANE_JAMMER.get(),
-                        (int) (property(DefaultProperties.EFFECT_DURATION) * ctx.modifiers().get(EBItems.DURATION_UPGRADE.get())),
+                        (int) (property(DefaultProperties.EFFECT_DURATION) * ctx.modifiers().get(SpellModifiers.DURATION)),
                         (int) (property(DefaultProperties.EFFECT_STRENGTH) * ctx.modifiers().get(SpellModifiers.POTENCY) - 1)));
             }
         }